requirements.txt
crawl4ai
crewai
crewai-tools
pandas
python-dotenv
opensky-api-1.3.0